+++ src/main/java/cn/staitech/annotation/vo/anno/AnnotationReq.java
@@ -20,5 +20,5 @@ public class AnnotationReq {
     private Long slideId;
 
     @ApiModelProperty(value = "脏器识别校对view页面必传参数：传递1")
-    private Long contourType;
+    private Integer contourType;
 }
